ag-library-manager/requirements.txt
python-dotenv==1.0.1
tqdm==4.66.1
